Timothy/Glance/config/f1_extensions/f1_api_wrappers/requirements.txt
fastapi
uvicorn
httpx
fastf1
pycountry
pytz
fastapi-cache2
svgwrite
numpy
